+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added receiving first result set on construct `internal/table/scanner.NewStream()`
 * Added experimental package `metrics` with SDK metrics
 * Fixed redundant trace call for finished `database/sql` transactions
 * Added repeater event type to wake-up func context